fix: green clippy gate — resolve all 31 lint errors across 28 files
Systematic cleanup: too_many_arguments allows on Bevy systems,
type_complexity allows on complex queries, dead_code allows on
contract types not yet wired, plus minor idiom fixes (collapsible_if,
is_none_or, derivable_impls, needless_borrow).

All gates pass: cargo check, cargo test (88/88), cargo clippy -D warnings (0 errors).

+++ dlc/pilot/src/economy/shop.rs
@@ -52,6 +52,7 @@ pub fn is_airport_exclusive(airport: AirportId, item_id: &str) -> bool {
     airport_premium_items(airport).iter().any(|(id, _)| *id == item_id)
 }
 
+#[allow(clippy::too_many_arguments)]
 pub fn handle_purchase(
     mut events: EventReader<PurchaseEvent>,
     mut gold: ResMut<Gold>,
@@ -150,7 +151,7 @@ pub fn restock_shop(
 
         for listing in active_shop.listings.iter_mut() {
             if let Some(stock) = &mut listing.stock {
-                let base_stock = 3 + ((day * 7 + listing.price) % 5) as u32;
+                let base_stock = 3 + (day * 7 + listing.price) % 5;
                 *stock = base_stock;
             }
         }
